Adding new entity event placement tools.
Now has a selection of entity event IDs, and modifiers are selected from
drop down lists once an entity type is selected.
Also added an exception catch in the main form's OnLoad - was mainly to
help debug entity definitions but I figure it's good to not proceed if
things aren't properly initialized.