Update grid drawing, room drawing, and scrollbars
Updated the tile and screen grid draw routines to no longer draw grids
outside the valid level coordinates.
Updated the room color and room bg draw routines to properly fill in the
full space for "12 tile tall" rooms.
Updated scrollbar handling to set the large change based on the view
size, and to try to minimize the amount of the display showing out of
bounds area.