Ensure at least some undrawn rooms are drawn.
Each time the view is redrawn, ensure a minimum of five rooms are
redrawn, if necessary, even if it goes over the time limit.  This
ensures it'll never get stuck in a state where it hits the time limit
just copying already-drawn rooms to the screen and never draws any new
rooms.