Round zoom values to one decimal place.
When zooming in/out using the track bar or the mouse wheel, the zoom
value will now be rounded to one decimal place to prevent floating point
error resulting in strings of many digits past the decimal point.  This
rounding does not occur if the user types in a zoom value directly.