Zoom on the center of the view.
The program will now try to keep the view centered where it was when
zooming in or out.  This is done by making the scroll positions floating
point (they should be used as int everywhere they already were
referenced), and calculating a new scroll position (with sub-tile
accuracy) to adjust the view and keep centered.